# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"GeneratedModelDTO": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"userId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"image",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GeneratedModelAttributes": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"image": {
						"type": "string"
					},
					"plyDirectory": {
						"type": "string"
					},
					"public": {
						"type": "boolean"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"userId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"image",
					"plyDirectory",
					"public",
					"date",
					"userId"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "api-reality-clone",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/server/ping": {
			"get": {
				"operationId": "Ping",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"timestamp": {
											"type": "string"
										},
										"status": {
											"type": "string"
										}
									},
									"required": [
										"timestamp",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Server"
				],
				"security": [],
				"parameters": []
			}
		},
		"/server/absolute-path": {
			"get": {
				"operationId": "AbsolutePath",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Server"
				],
				"security": [],
				"parameters": []
			}
		},
		"/image/compute-gaussian": {
			"post": {
				"operationId": "ComputeGaussian",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Image"
				],
				"security": [],
				"parameters": []
			}
		},
		"/image": {
			"get": {
				"operationId": "GetGeneratedModel",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/GeneratedModelDTO"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Image"
				],
				"security": [],
				"parameters": []
			}
		},
		"/image/{id}": {
			"get": {
				"operationId": "GetGaussianById",
				"responses": {
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Image"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/image/delete/{id}": {
			"delete": {
				"operationId": "DeleteGeneratedModel",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Image"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/image/download/{id}": {
			"get": {
				"operationId": "DownloadGaussianById",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Image"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/image/add-gaussian": {
			"post": {
				"operationId": "AddGaussian",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Image"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GeneratedModelAttributes"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}